Doxygen: Fix Many Warnings
Fix warnings and refactor code to build a clean doxygen
documentation.
