+++ src/IO/ADIOS/ParallelADIOS1IOHandler.cpp
@@ -319,9 +319,11 @@ ParallelADIOS1IOHandlerImpl::initialize_group(std::string const &name)
     return group;
 }
 
+#ifndef DOXYGEN_SHOULD_SKIP_THIS
 #define CommonADIOS1IOHandlerImpl ParallelADIOS1IOHandlerImpl
 #include "CommonADIOS1IOHandler.cpp"
 #undef CommonADIOS1IOHandlerImpl
+#endif
 
 #else
 #   if openPMD_HAVE_MPI
